requirements.txt
Django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
gunicorn==20.1.0
drf-yasg==1.21.7
